Refactor: Navigate to LoginFragment after successful account creation
This commit updates the `CreateAccountController` to automatically navigate the user to the `LoginFragment` after successfully creating either a student or business account.

Previously, after account creation, the user remained on the account creation screen. Now, `handleCreateStudentAccountCancel()` and `handleCreateBusinessAccountCancel()` are called, respectively, after displaying the success Toast message. These methods are responsible for navigating to the `LoginFragment`.