Refactor: Initialize loading spinner in LoginActivity
The `loadingSpinner` ConstraintLayout, which contains a ProgressBar, is now initialized in the `onCreate` method of `LoginActivity.java`. This was done by finding the view by its ID `R.id.login_loading_spinner`.

The `activity_login.xml` for `sw600dp-land` has been updated to include this `login_loading_spinner` ConstraintLayout. It is configured to cover the entire screen, has a dim background, and centers a ProgressBar. Its visibility is initially set to `gone`.